+++ app/policies/feedback_policy.rb
@@ -1,0 +1,15 @@
+class FeedbackPolicy < ApplicationPolicy
+  class Scope < Scope
+    def resolve
+      scope.all
+    end
+  end
+
+  def new?
+    true
+  end
+
+  def create?
+    true
+  end
+end
