--- app/views/feedback/new.html.erb
@@ -1,1 +1,0 @@
-<h1>Feedback View</h1>
