+++ app/views/shared/_footer.html.erb
@@ -14,7 +14,7 @@
     </div>
   </div>
   <div class="tab-pane fade <%= 'show active' if @tab == "feedback" %>" id="pills-mobile-feedback" role="tabpanel" aria-labelledby="pills-mobile-feedback-tab">
-    <p>FEEDBACK</p>
+    <%= render 'feedback/new', event: event %>
   </div>
 </div>
 <ul class="nav nav-pills mt-3 justify-content-center" id="pills-tab" role="tablist">
