+++ test/policies/feedback_policy_test.rb
@@ -1,0 +1,18 @@
+require 'test_helper'
+
+class FeedbackPolicyTest < ActiveSupport::TestCase
+  def test_scope
+  end
+
+  def test_show
+  end
+
+  def test_create
+  end
+
+  def test_update
+  end
+
+  def test_destroy
+  end
+end
